update form and error handling
worked on the update form, made comments on what i need to do, built out form UI, added skeleton for the logic.

attempted to play around with error handling for when you try to search for an id that doesn't exist. I can console.log the error im getting, but not sure how to use that yet. Maybe i need to define the error as a type. Will need to research more.