pharmacy and updateForm
I extracted the pharmacy component away from everything else. It is still currently working on a button to load pharmacies. Next step is to pre-fetch this information.

Each pharmacy displayed has an edit button which currently sets the pharmacy state to the target pharmacy. Need to have this button open up the updateForm component.

updateForm should be able to access this pharmacy state now