added utils to state & formatted the pharmacy list and search view
I added a util state to the store, I believe it is loose enough that a variety of different types of utility states could be added to it. For now I have added a viewAll state which indicates if the user wants to view all of the pharmacies or not. This is so the search and the view all button can both reach this state and toggle it, thereby toggling how many pharmacies are rendered. There might be an easier way to do this but for now it keeps my code seperated. I also formated the view for pharmacyList, it includes conditional rendering which I can hopefully refactor out later.